+++ test/index_test.rb
@@ -112,6 +112,9 @@ class IndexTest < Minitest::Test
   end
 
   def test_analyzed_only
+    # skip for 5.0 since it throws
+    # Cannot search on field [alt_description] since it is not indexed.
+    skip unless elasticsearch_below50?
     store [{name: "Product A", alt_description: "Hello"}]
     assert_search "*", [], where: {alt_description: "Hello"}
   end