+++ test/query_test.rb
@@ -12,8 +12,9 @@ class QueryTest < Minitest::Test
   end
 
   def test_with_effective_min_score
+    min_score = elasticsearch_below50? ? 0.1 : 1
     store_names ["Milk", "Milk2"]
-    assert_search "milk", ["Milk"], body_options: {min_score: 0.1}
+    assert_search "milk", ["Milk"], body_options: {min_score: min_score}
   end
 
   def test_with_uneffective_min_score