+++ test/sql_test.rb
@@ -163,6 +163,7 @@ class SqlTest < Minitest::Test
   end
 
   def test_select_v2_include
+    skip unless elasticsearch_below50?
     store [{name: "Product A", user_ids: [1, 2]}]
     result = Product.search("product", load: false, select_v2: {include: [:name]}).first
     assert_equal %w(id name), result.keys.reject { |k| k.start_with?("_") }.sort
@@ -171,6 +172,7 @@ class SqlTest < Minitest::Test
   end
 
   def test_select_v2_exclude
+    skip unless elasticsearch_below50?
     store [{name: "Product A", user_ids: [1, 2], store_id: 1}]
     result = Product.search("product", load: false, select_v2: {exclude: [:name]}).first
     assert_nil result.name
@@ -179,9 +181,38 @@ class SqlTest < Minitest::Test
   end
 
   def test_select_v2_include_and_exclude
+    skip unless elasticsearch_below50?
     # let's take this to the next level
     store [{name: "Product A", user_ids: [1, 2], store_id: 1}]
     result = Product.search("product", load: false, select_v2: {include: [:store_id], exclude: [:name]}).first
+    assert_equal 1, result.store_id
+    assert_nil result.name
+    assert_nil result.user_ids
+  end
+
+  def test_select_v2_includes
+    skip if elasticsearch_below50?
+    store [{name: "Product A", user_ids: [1, 2]}]
+    result = Product.search("product", load: false, select_v2: {includes: [:name]}).first
+    assert_equal %w(id name), result.keys.reject { |k| k.start_with?("_") }.sort
+    assert_equal "Product A", result.name
+    assert_nil result.store_id
+  end
+
+  def test_select_v2_excludes
+    skip if elasticsearch_below50?
+    store [{name: "Product A", user_ids: [1, 2], store_id: 1}]
+    result = Product.search("product", load: false, select_v2: {excludes: [:name]}).first
+    assert_nil result.name
+    assert_equal [1, 2], result.user_ids
+    assert_equal 1, result.store_id
+  end
+
+  def test_select_v2_include_and_excludes
+    skip if elasticsearch_below50?
+    # let's take this to the next level
+    store [{name: "Product A", user_ids: [1, 2], store_id: 1}]
+    result = Product.search("product", load: false, select_v2: {includes: [:store_id], excludes: [:name]}).first
     assert_equal 1, result.store_id
     assert_nil result.name
     assert_nil result.user_ids