+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.3.5 [unreleased]
 
+- Added support for Elasticsearch 5.0 beta
 - Added `request_params` option
 - Added `filterable` option